requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.2.2
faiss-cpu==1.11.0
fastapi==0.115.12
fastapi-cli==0.0.7
faust==1.10.4
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
nomic-embed-text==0.2.0
numpy==2.0.2
orjson==3.10.18
packaging==25.0
passlib==1.7.4
pydantic==2.11.4
pydantic-extra-types==2.10.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.1
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.32.3
rich==14.0.0
rich-toolkit==0.14.5
shellingham==1.5.4
sniffio==1.3.1
starlette==0.46.2
typer==0.15.3
typing-inspection==0.4.0
typing_extensions==4.13.2
ujson==5.10.0
urllib3==2.4.0
uvicorn==0.34.2
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
zipp==3.21.0
